requirements.txt
Cython==0.27.3
h5py==2.7.1
jupyter==1.0.0
Keras==2.0.5
numpy==1.14.1
pandas==0.22.0
scikit-learn==0.19.1
scipy==1.0.0
#tensorflow==1.6.0
tensorflow-gpu==1.6.0
